Break dependence on MT-specific stuff in ImageUtil
And enforce naming standards (FLIP_DIRECTION -> FlipDirection)
